# Create the query only once