--- a/topiary/benches/benchmark.rs
+++ b/topiary/benches/benchmark.rs
@@ -2,15 +2,16 @@ use criterion::async_executor::FuturesExecutor;
 use criterion::{criterion_group, criterion_main, Criterion};
 use std::fs;
 use std::io;
-use topiary::Configuration;
 use topiary::{formatter, Operation};
+use topiary::{Configuration, TopiaryQuery};
 
 async fn format() {
     let input = fs::read_to_string("tests/samples/input/ocaml.ml").unwrap();
-    let query = fs::read_to_string("../languages/ocaml.scm").unwrap();
+    let query_content = fs::read_to_string("../languages/ocaml.scm").unwrap();
     let configuration = Configuration::parse_default_configuration().unwrap();
     let language = configuration.get_language("ocaml").unwrap();
     let grammar = language.grammar().await.unwrap();
+    let query = TopiaryQuery::new(&grammar, &query_content).unwrap();
 
     let mut input = input.as_bytes();
     let mut output = io::BufWriter::new(Vec::new());