The query is then used between the formatting and the idempotence check. The TopiaryQuery struct helps keep the API clean.
### Benchmark comparison with `skip_idempotence` set to false.
#### main:
```
format_ocaml            time:   [655.12 ms 658.89 ms 662.75 ms]
```

#### after this PR:
```
format_ocaml            time:   [430.05 ms 432.33 ms 434.68 ms]
                        change: [-34.911% -34.385% -33.879%] (p = 0.00 < 0.05)
```


### Flamegraphs of a simple OCaml format
#### main:
![flamegraph_main](https://github.com/tweag/topiary/assets/10973664/0752a307-4c67-4941-b882-c637df994f5e)

#### after this PR:
![flamegraph_new](https://github.com/tweag/topiary/assets/10973664/3faa0a47-dfae-4728-b00f-fc28ebd9d30c)